src/chapter-02/entra-api/requirements.txt
Flask
azure-identity
openai
gunicorn    